code box properly positions between top/bottom and left/right of cell

+++ quadratic-core/src/grid/js_types.rs
@@ -201,8 +201,8 @@ pub struct JsCodeCell {
 #[derive(Serialize, Deserialize, Debug, Clone)]
 #[cfg_attr(feature = "js", derive(ts_rs::TS))]
 pub struct JsRenderCodeCell {
-    pub x: i64,
-    pub y: i64,
+    pub x: i32,
+    pub y: i32,
     pub w: u32,
     pub h: u32,
     pub language: CodeCellLanguage,
